refactor: Move unregister function to user from auth
