+++ src/user/service.py
@@ -1,17 +1,25 @@
 from sqlalchemy.orm import Session
 
-from auth.util.redis import get_token_from_redis
+from auth.service import REFRESH_TOKEN_REDIS
+from auth.util.redis import delete_token_from_redis, get_token_from_redis
 from src.models import User
-from src.response.error_definitions import UserAlreadyExist, UserNotFound
-from user import repository
-from user.schemas import UserReq, UserRes
+from src.project import repository as project_repository
+from src.project_user import repository as project_user_repository
+from src.response.error_definitions import (
+    ProjectUserExist,
+    UserAlreadyExist,
+    UserNotFound,
+)
+from src.user import repository as user_repository
+from src.user.schemas import UserReq, UserRes
+from src.user_repository import repository as user_repository_repository
 
 
 async def create_user(db: Session, user_req: UserReq):
     """
     Create new user
     """
-    existing_user = repository.find_user_by_github_id(db, user_req.github_id)
+    existing_user = user_repository.find_user_by_github_id(db, user_req.github_id)
     if existing_user:
         raise UserAlreadyExist()
 
@@ -28,7 +36,7 @@ async def create_user(db: Session, user_req: UserReq):
         profile_img=user_req.profile_img,
     )
 
-    saved_user = repository.create_user(db, new_user)
+    saved_user = user_repository.create_user(db, new_user)
     return saved_user
 
 
@@ -36,7 +44,7 @@ async def update_user(db: Session, user_req: UserReq):
     """
     Update existing user
     """
-    existing_user = repository.find_user_by_github_id(db, user_req.github_id)
+    existing_user = user_repository.find_user_by_github_id(db, user_req.github_id)
     if not existing_user:
         raise UserNotFound()
 
@@ -47,7 +55,7 @@ async def update_user(db: Session, user_req: UserReq):
     existing_user.career = user_req.career
     existing_user.profile_img = user_req.profile_img
 
-    updated_user = repository.update_user(db, existing_user)
+    updated_user = user_repository.update_user(db, existing_user)
     return updated_user
 
 
@@ -55,7 +63,7 @@ async def update_github_user(db: Session, github_id: str):
     """
     Update existing user info with GitHub info
     """
-    existing_user = repository.find_user_by_github_id(db, github_id)
+    existing_user = user_repository.find_user_by_github_id(db, github_id)
     if not existing_user:
         raise UserNotFound()
 
@@ -72,3 +80,28 @@ def search_users_by_name(user_name: str, db: Session):
         user_res_list.append(user_res)
 
     return user_res_list
+
+
+async def unregister(user_id: int, db: Session):
+    """
+    Unregister user and clean up all related data
+    """
+    # Check if user has project or owns project
+    project_list = project_user_repository.find_all_projects_by_user_id(db, user_id)
+    owned_project_list = project_repository.find_project_by_owner(db, user_id)
+    if project_list or owned_project_list:
+        raise ProjectUserExist()
+
+    # Delete user repositories
+    user_repository_repository.delete_all_repositories_by_user_id(db, user_id)
+
+    # Delete user from database
+    user = user_repository.find_user_by_user_id(db, user_id)
+    if not user:
+        raise UserNotFound()
+
+    user_repository.delete_user(db, user)
+
+    # Delete tokens from redis
+    redis_refresh_token = await get_token_from_redis(REFRESH_TOKEN_REDIS, user_id)
+    await delete_token_from_redis(redis_refresh_token)
